Remove all references to posixpath from ABCs
